requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
gunicorn==21.2.0
